# contracts/identity.openapi.yaml
openapi: 3.0.3
info:
  title: Zennyt - Identity API
  description: >
    Identity bounded context: local, Google, and Apple authentication, JWT lifecycle,
    candidate/student and recruiter onboarding, and professional profiles.
  version: 2.0.0
servers:
  - url: https://api.zennyt.com/api/v1
    description: Production
  - url: https://api-staging.zennyt.com/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    put:
      operationId: putUsersMe
      tags: [Users]
      summary: Update common account fields
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: './common.openapi.yaml#/components/responses/BadRequest' }
    delete:
      operationId: deleteUsersMe
      tags: [Users]
      summary: Delete the current account (soft delete and anonymization)
      responses:
        '204': { description: Account deleted }
        '401': { $ref: './common.openapi.yaml#/components/responses/Unauthorized' }
  /users/me/avatar:
    post:
      operationId: postUsersMeAvatar
      tags: [Users]
      summary: Upload (or replace) the current account avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG, JPEG or WEBP, max 5MB).
      responses:
        '200':
          description: Avatar uploaded; account with profileImageUrl set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: './common.openapi.yaml#/components/responses/BadRequest' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    delete:
      operationId: deleteUsersMeAvatar
      tags: [Users]
      summary: Remove the current account avatar
      responses:
        '200':
          description: Avatar removed; account with profileImageUrl cleared
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
  /users/me/deactivate:
    post:
      operationId: postUsersMeDeactivate
      tags: [Users]
      summary: Deactivate the current account (reversible at next login)
      responses:
        '204': { description: Account deactivated }
        '401': { $ref: './common.openapi.yaml#/components/responses/Unauthorized' }
  /users/me/role:
    patch:
      operationId: patchUsersMeRole
      tags: [Users]
      summary: Change the account role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Role changed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: './common.openapi.yaml#/components/responses/BadRequest' }
...
  /profiles:
    post:
      operationId: postProfiles
      tags: [Profiles]
      summary: Create the current candidate or student profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileInput' }
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '403': { $ref: './common.openapi.yaml#/components/responses/Forbidden' }
        '409': { $ref: './common.openapi.yaml#/components/responses/Conflict' }
  /profiles/me:
    get:
      operationId: getProfilesMe
      tags: [Profiles]
      summary: Get the current professional profile
      responses:
        '200':
          description: Current profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    put:
      operationId: putProfilesMe
      tags: [Profiles]
      summary: Update the current professional profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileInput' }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
  /profiles/me/cv:
    post:
      operationId: postProfilesMeCv
      tags: [Profiles]
      summary: Upload (or replace) the current profile's CV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CV file (PDF, DOC or DOCX, max 5MB).
      responses:
        '200':
          description: CV uploaded; profile with cvUrl set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: './common.openapi.yaml#/components/responses/BadRequest' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    delete:
      operationId: deleteProfilesMeCv
      tags: [Profiles]
      summary: Remove the current profile's CV
      responses:
        '200':
          description: CV removed; profile with cvUrl cleared
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
...
  /profiles/{profileId}:
    get:
      operationId: getProfilesByProfileid
      tags: [Profiles]
      summary: Get a professional profile by ID
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
  /profiles/me/skills:
    get:
      operationId: getProfilesMeSkills
      tags: [Profiles]
      summary: List skills, optionally filtered by type
      parameters:
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/SkillType' }
      responses:
        '200':
          description: Skills
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Skill' }
    post:
      operationId: postProfilesMeSkills
      tags: [Profiles]
      summary: Add a skill
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkillInput' }
      responses:
        '201':
          description: Skill added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
  /profiles/me/skills/{id}:
    put:
      operationId: putProfilesMeSkillsById
      tags: [Profiles]
      summary: Update a skill
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkillInput' }
      responses:
        '200':
          description: Skill updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
    delete:
      operationId: deleteProfilesMeSkillsById
      tags: [Profiles]
      summary: Delete a skill
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204': { description: Skill deleted }
  /profiles/me/positions:
    get:
      operationId: getProfilesMePositions
      tags: [Profiles]
      summary: List positions
      responses:
        '200':
          description: Positions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Position' }
    post:
      operationId: postProfilesMePositions
      tags: [Profiles]
      summary: Add a position
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionInput' }
      responses:
        '201':
          description: Position added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
  /profiles/me/positions/{id}:
    put:
      operationId: putProfilesMePositionsById
      tags: [Profiles]
      summary: Update a position
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionInput' }
      responses:
        '200':
          description: Position updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
    delete:
      operationId: deleteProfilesMePositionsById
      tags: [Profiles]
      summary: Delete a position
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204': { description: Position deleted }
  /profiles/me/certifications:
    get:
      operationId: getProfilesMeCertifications
      tags: [Profiles]
      summary: List certifications
      responses:
        '200':
          description: Certifications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Certification' }
    post:
      operationId: postProfilesMeCertifications
      tags: [Profiles]
      summary: Add a certification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CertificationInput' }
      responses:
        '201':
          description: Certification added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Certification' }
  /profiles/me/certifications/{id}:
    put:
      operationId: putProfilesMeCertificationsById
      tags: [Profiles]
      summary: Update a certification
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CertificationInput' }
      responses:
        '200':
          description: Certification updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Certification' }
    delete:
      operationId: deleteProfilesMeCertificationsById
      tags: [Profiles]
      summary: Delete a certification
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204': { description: Certification deleted }
  /profiles/me/education:
    get:
      operationId: getProfilesMeEducation
      tags: [Profiles]
      summary: List education records
      responses:
        '200':
          description: Education records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Education' }
    post:
      operationId: postProfilesMeEducation
      tags: [Profiles]
      summary: Add an education record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EducationInput' }
      responses:
        '201':
          description: Education record added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Education' }
  /profiles/me/education/{id}:
    put:
      operationId: putProfilesMeEducationById
      tags: [Profiles]
      summary: Update an education record
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EducationInput' }
      responses:
        '200':
          description: Education record updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Education' }
    delete:
      operationId: deleteProfilesMeEducationById
      tags: [Profiles]
      summary: Delete an education record
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204': { description: Education record deleted }
components:
  securitySchemes:
    bearerAuth: { $ref: './common.openapi.yaml#/components/securitySchemes/bearerAuth' }
  parameters:
    ProfileId:
      name: profileId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    ItemId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
  schemas:
    Role:
      type: string
      enum: [CANDIDATE, STUDENT, RECRUITER, ADMIN]
...
    UserUpdate:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string, maxLength: 100 }
        lastName: { type: string, maxLength: 100 }
        phoneNumber: { type: string, maxLength: 30 }
        city: { type: string, maxLength: 100 }
        country: { type: string, maxLength: 100 }
        address: { type: string, maxLength: 255 }
    User:
      allOf:
        - $ref: '#/components/schemas/UserUpdate'
        - type: object
          required: [id, email, role, termsAccepted, emailVerified, active, createdAt, updatedAt]
          properties:
            id: { type: integer, format: int64 }
            email: { type: string, format: email }
            role: { $ref: '#/components/schemas/Role' }
            # L'avatar est en lecture seule : il se gère via POST/DELETE /users/me/avatar.
            profileImageUrl: { type: string, maxLength: 500, readOnly: true }
            termsAccepted: { type: boolean }
            emailVerified: { type: boolean }
            active: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    ProfileInput:
      type: object
      properties:
        currentPosition: { type: string, maxLength: 150 }
        lookingFor: { type: string, maxLength: 150 }
        workplaceType:
          type: string
          enum: [ONSITE, REMOTE, HYBRID, FLEXIBLE]
        jobType:
          type: string
          enum: [FULL_TIME, PART_TIME, INTERNSHIP, FREELANCE, CONTRACT]
        targetJobLocation: { type: string, maxLength: 150 }
        yearsOfExperience: { type: integer, minimum: 0 }
        softSkillsScore: { type: integer, minimum: 0, maximum: 100 }
        aboutMe: { type: string }
        openInternationally: { type: boolean, default: false }
        availabilityType:
          type: string
          enum: [IMMEDIATELY, SELECT_DATE]
        availabilityDate: { type: string, format: date }
        resumeAiUrl: { type: string, maxLength: 500 }
        portfolioUrl: { type: string, maxLength: 500 }
    Profile:
      allOf:
        - $ref: '#/components/schemas/ProfileInput'
        - $ref: '#/components/schemas/OwnedResource'
        - type: object
          properties:
            cvUrl:
              type: string
              maxLength: 500
              readOnly: true
              description: URL of the uploaded CV (set via POST /profiles/me/cv).
            skills:
              type: array
              items: { $ref: '#/components/schemas/Skill' }
            positions:
              type: array
              items: { $ref: '#/components/schemas/Position' }
            certifications:
              type: array
              items: { $ref: '#/components/schemas/Certification' }
            education:
              type: array
              items: { $ref: '#/components/schemas/Education' }
    SkillType:
      type: string
      enum: [TECHNICAL, SOFT]
    SkillInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string, maxLength: 100 }
        type: { $ref: '#/components/schemas/SkillType' }
        level: { type: integer, minimum: 1, maximum: 5 }
    Skill:
      allOf:
        - $ref: '#/components/schemas/SkillInput'
        - $ref: '#/components/schemas/ProfileItem'
    PositionInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 150 }
        companyName: { type: string, maxLength: 150 }
        location: { type: string, maxLength: 150 }
        description: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        current: { type: boolean, default: false }
    Position:
      allOf:
        - $ref: '#/components/schemas/PositionInput'
        - $ref: '#/components/schemas/ProfileItem'
    CertificationInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 150 }
        issuer: { type: string, maxLength: 150 }
        completionDate: { type: string, format: date }
        credentialId: { type: string, maxLength: 150 }
        credentialUrl: { type: string, maxLength: 500 }
    Certification:
      allOf:
        - $ref: '#/components/schemas/CertificationInput'
        - $ref: '#/components/schemas/ProfileItem'
    EducationInput:
      type: object
      required: [degree]
      properties:
        degree: { type: string, maxLength: 150 }
        school: { type: string, maxLength: 150 }
        fieldOfStudy: { type: string, maxLength: 150 }
        description: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    Education:
      allOf:
        - $ref: '#/components/schemas/EducationInput'
        - $ref: '#/components/schemas/ProfileItem'
    OwnedResource:
      type: object
      required: [id, userId, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProfileItem:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...